Move `waitUntilScoreColumnIsLoaded()`
The function must be moved from _src/util.ts_ to _src/tasks/utils.ts_ as the related imports for this function causes Karma testing errors that `document is Undefined`.

Background: Some function from _src/util.ts_ are imported in _Measures.ts_. The measurments are unit tested by karma.

The root cause is `_init()` function in phovea_core/src/index.ts, which requires a `document` and is automatically called.
