+++ src/util.ts
@@ -1,10 +1,6 @@
 import {IMeasureResult} from './interfaces';
 import {IColumnDesc} from 'lineupjs';
-import {IServerColumn} from 'tdp_core/src/rest';
-import {RankingAdapter} from './RankingAdapter';
-import {uniqueId} from 'phovea_core/src';
 
-const EVENT_DATA_LOADED = 'dataLoaded';
 /**
  * Returns:
  *  intersection: elements in both items
@@ -173,33 +169,4 @@ export function shuffle(arr: any[]): any[] {
  */
 export function isScoreColumn(colDesc: IColumnDesc) {
   return colDesc.hasOwnProperty('lazyLoaded');
-}
-
-
-/**
- * This functions returns a promise that gets resolved, once the score column is loaded.
- * The notification is implemented based on a flag or an event.
- */
-export function waitUntilScoreColumnIsLoaded(ranking: RankingAdapter, desc: IColumnDesc): Promise<any> {
-  const scoreColumn = (<any[]>ranking.getScoreColumns()).find((col) => (<IServerColumn>col.desc).column === (<IServerColumn>desc).column);
-
-  if(!scoreColumn) {
-    return Promise.resolve();
-  }
-
-  return new Promise((resolve) => {
-    if(scoreColumn.isLoaded()) {
-      // console.log('data already loaded for', scoreColumn.desc.label);
-      resolve();
-      return;
-    }
-
-    const uniqueSuffix = `.tourdino${uniqueId()}`;
-
-    scoreColumn.on(EVENT_DATA_LOADED + uniqueSuffix, () => { // add suffix with unique Id to resolve all promises for each instance of scoreColumn
-      scoreColumn.on(EVENT_DATA_LOADED + uniqueSuffix, null);
-      // console.log('data loaded (notified by event) for', scoreColumn.desc.label);
-      resolve();
-    });
-  });
 }
